requirements.txt
langchain-openai
pandas~=1.5.3
numpy~=1.26.4
sqlalchemy~=2.0.38
pymysql
python-dotenv~=1.0.1
requests~=2.32.3
uvicorn==0.27.0
gunicorn==21.2.0
openai
pydantic~=2.10.6
fastapi==0.109.0
langchain
langchain-anthropic
langchain-community
pinecone~=6.0.1
langchain_core
langchain_experimental
sqlalchemy
